// src/api/generated/models/devProfileResource.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * brew
 * OpenAPI spec version: 0.0.1
 */
import type { UserResource } from './userResource';
import type { DevProfileResourceBioPt } from './devProfileResourceBioPt';
import type { DevProfileResourceBioEn } from './devProfileResourceBioEn';
import type { DevProfileResourceTranslationStatusLabel } from './devProfileResourceTranslationStatusLabel';
import type { EmploymentHistoryResource } from './employmentHistoryResource';
import type { ProjectHistoryResource } from './projectHistoryResource';
import type { AcademicBackgroundResource } from './academicBackgroundResource';
import type { AdditionalCourseResource } from './additionalCourseResource';
import type { DevSoftSkillResource } from './devSoftSkillResource';
import type { AddressResource } from './addressResource';
import type { DevProfileResourceCreatedAt } from './devProfileResourceCreatedAt';
import type { DevProfileResourceUpdatedAt } from './devProfileResourceUpdatedAt';

export interface DevProfileResource {
  id: string;
  user_id: string;
  user?: UserResource;
  name: string;
  bio: string;
  bio_pt: DevProfileResourceBioPt;
  bio_en: DevProfileResourceBioEn;
  translation_status: string;
  translation_status_label: DevProfileResourceTranslationStatusLabel;
  cpf: string;
  phone: string;
  seniority_level: string;
  seniority_level_label: string;
  seniority_tested: boolean;
  specialty: string;
  specialty_label: string;
  birthdate: string;
  score: number;
  open_to_work: boolean;
  open_to_relocation: boolean;
  is_blocked: boolean;
  employment_histories?: EmploymentHistoryResource[];
  project_histories?: ProjectHistoryResource[];
  academic_backgrounds?: AcademicBackgroundResource[];
  additional_courses?: AdditionalCourseResource[];
  dev_soft_skills?: DevSoftSkillResource[];
  address?: AddressResource;
  created_at: DevProfileResourceCreatedAt;
  updated_at: DevProfileResourceUpdatedAt;
}